[move]Rankクラスをlogbook.constantsパッケージへ移動
