+++ src/main/java/logbook/internal/gui/CalcExpController.java
@@ -40,8 +40,8 @@ import logbook.bean.Ship;
 import logbook.bean.ShipCollection;
 import logbook.bean.ShipMst;
 import logbook.constants.ExpTable;
+import logbook.constants.Rank;
 import logbook.internal.LoggerHolder;
-import logbook.internal.Rank;
 import logbook.internal.Ships;
 
 public class CalcExpController extends WindowController {
